move model location and refactor database.
