+++ service_notification/src/main/kotlin/org/thechance/service_notification/domain/model/User.kt
@@ -1,4 +1,4 @@
-package org.thechance.service_notification.domain.entity
+package org.thechance.service_notification.domain.model
 
 data class User (
     val id: String,
